% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comperimeter.R
\name{comp.flex}
\alias{comp.flex}
\title{Create comperimeter tables}
\usage{
comp.flex(form, data, param = TRUE, m.deci = 1, p.deci = 1, lang = "hr")
}
\arguments{
\item{form}{formula specified as 1 ~ var1 + var2 +varX}

\item{data}{dataset}

\item{param}{should parametrics (M and SD) be included (default=TRUE)}

\item{m.deci}{number of decimal points for parametrics (default=1)}

\item{p.deci}{number of decimal points for percentages (default=1)}

\item{lang}{language (defaut="hr")}
}
\value{
data.frame with results
}
\description{
Function creates basic comperimeter flextable with optional descriptives. Variables must be labelled.
}
