% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{freq.flex}
\alias{freq.flex}
\title{Frequencies table}
\usage{
freq.flex(vari, data, deci = 1)
}
\arguments{
\item{vari}{variable name as string}

\item{data}{dataset}

\item{deci}{number of decimals for percentages (default=1)}
}
\value{
Flextable object
}
\description{
Frequencies table
}
