% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{freq.flex.rmd}
\alias{freq.flex.rmd}
\title{The most simple frequencies}
\usage{
freq.flex.rmd(vari, data, deci = 1, heading = 2)
}
\arguments{
\item{vari}{variable name as string}

\item{data}{dataset}

\item{deci}{number of decimals for percentages (default=1)}

\item{heading}{Which heading should be used for title (default=2)}
}
\value{
Title and table
}
\description{
The most simple frequencies
}
