% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load.SPSS}
\alias{load.SPSS}
\title{Load dataset from SPSS}
\usage{
load.SPSS(path = "file.choose()")
}
\arguments{
\item{path}{path to dataset (default = file.choose())}
}
\value{
loaded data.frame
}
\description{
Function enables loading from SPSS by using haven's method
}
