% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives.R, R/simple_descriptives2.R
\name{r.kurtosis.se}
\alias{r.kurtosis.se}
\title{Calculates standard error for excess kurtosis of the variable}
\usage{
r.kurtosis.se(target_variable)

r.kurtosis.se(target_variable)
}
\arguments{
\item{target_variable}{numeric variable}
}
\value{
SE for excess kurtosis

SE for excess kurtosis
}
\description{
uses SPSS Book of algorithms v20

uses SPSS Book of algorithms v20
}
