% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rajter.flex-package.R
\docType{package}
\name{rajter.flex-package}
\alias{rajter.flex}
\alias{rajter.flex-package}
\title{rajter.flex: A Flextable Based Statistics Package}
\description{
The package for doing statistics and creating pretty word tables using flaxtable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrajter/rajter.flex}
  \item Report bugs at \url{https://github.com/mrajter/rajter.flex/issues}
}

}
\author{
\strong{Maintainer}: Miroslav Rajter \email{miroslav.rajter@gmail.com} (\href{https://orcid.org/0000-0002-2608-3196}{ORCID})

}
\keyword{internal}
