% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest.flex}
\alias{ttest.flex}
\title{Independent samples t-test}
\usage{
ttest.flex(
  data,
  form,
  effect = "d",
  type = "two.sided",
  deci = 2,
  conf.level = 0.95,
  lang = "hr"
)
}
\arguments{
\item{data}{dataset}

\item{form}{formula in format dependent ~ independent. Can have multiple dependent varibles (dep1+dep2+dep3~ind) or multiple independent variables (dep~ind1+ind2+ind3), but NOT BOTH.}

\item{effect}{effect size "d" (default) for Cohen's d or "prob_sup" for probability of superiority.}

\item{type}{type of test (one or two sided). Options are "two.sided" (default), "greater" and "less".}

\item{deci}{number of decimal places for M}

\item{conf.level}{confidence level}

\item{lang}{language of the output. Options are "hr" (default) and "en".}
}
\value{
list with table as flextable and data frame with results
}
\description{
Independent samples t-test
}
