% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_descriptives.R
\name{des.flex}
\alias{des.flex}
\title{Provide basic descriptives}
\usage{
des.flex(data, vars, param_set = "standard", by = FALSE, deci = 1, lang = "hr")
}
\arguments{
\item{data}{data.frame with variables}

\item{vars}{variables as list as strings}

\item{param_set}{list of parameters with order in the table (as strings):
\itemize{
\item default - "standard"
\item "simple" - N, Min, Max, M, SD
\item "standard" - N, M, SD, Min, Q1, C, Q3, Max, Skew, Kurt, SW, SWp
\item other available statistics - missing valuse(Miss), Kolmogorov-Smirnov (KS, KSp), standard errors for Skewness and Kurtosis (SEskew, SEkurt)
}
default is (N, M, SD, Min, Q1, C, Q3, Max). Optional parametres include missing (miss), variance (Var), Skewness (Skew), SE Skewnes (SE_skew), Kurtosis (Kurt), SE Kurtosis (SE_kurt), K-S test with Lilliefors correction (KS), K-S significance (KSp), Shapiro-Wilk test (SW), S-W significance (SWp)}

\item{by}{section variable, default=FALSE}

\item{deci}{number of decimals for M, everything else is done automatically (default=1)}

\item{lang}{language (default is "hr")}
}
\value{
data.frame with results
}
\description{
Simple descriptive statistics
}
